+++ src/reshape.jl
@@ -29,18 +29,6 @@ function stack(t::D, by = pkeynames(t); select = isa(t, NDSparse) ? valuenames(t
     convert(collectiontype(D), Columns(bycols), Columns(labelcol, valuecol, names = [variable, value]))
 end
 
-function unstack(::Type{D}, ::Type{T}, key, val, cols::AbstractVector{S}) where {D <:Dataset, T, S}
-    dest_val = Columns((DataValues.DataValueArray{T}(length(val)) for i in cols)...; names = cols)
-    for (i, el) in enumerate(val)
-        for j in el
-            k, v = j
-            isna(columns(dest_val, S(k))[i]) || error("Repeated values with same label are not allowed")
-            columns(dest_val, S(k))[i] = v
-        end
-    end
-    convert(collectiontype(D), key, dest_val)
-end
-
 """
     unstack(t, by = pkeynames(t); variable = :variable, value = :value)
 
@@ -61,5 +49,16 @@ function unstack(t::D, by = pkeynames(t); variable = :variable, value = :value) 
     S = eltype(colnames(t))
     cols = S.(union(columns(t, variable)))
     T = eltype(columns(t, value))
-    unstack(D, T isa Type{<:DataValue} ? eltype(T) : T, pkeys(tgrp), columns(tgrp, value), cols)
+    unstack(D, Base.nonmissingtype(T), pkeys(tgrp), columns(tgrp, value), cols)
+end
+
+function unstack(::Type{D}, ::Type{T}, key, val, cols::AbstractVector{S}) where {D <:Dataset, T, S}
+    dest_val = Columns((Array{Union{T, Missing}}(undef, length(val)) for i in cols)...; names = cols)
+    for (i, el) in enumerate(val)
+        for (k, v) in el
+            ismissing(columns(dest_val, S(k))[i]) || error("Repeated values with same label are not allowed")
+            columns(dest_val, S(k))[i] = v
+        end
+    end
+    convert(collectiontype(D), key, dest_val)
 end