+++ test/test_utils.jl
@@ -1,5 +1,3 @@
-using Test, IndexedTables, Serialization
-
 let a = NDSparse([12,21,32], [52,41,34], [11,53,150]), b = NDSparse([12,23,32], [52,43,34], [56,13,10])
     p = vec(collect(IndexedTables.product(a, b)))
     @test p == [(11,56), (11,13), (11,10), (53,56), (53,13), (53,10), (150,56), (150,13), (150,10)]